// src/interfaces/HpCharInt.ts
export interface HpCharInt {
  _id: string;
  name: string;
  role: string;
  house: string;
  school: string;
  wand: string;
  boggart: string;
  patronus: string;
  __v: number;
  ministryOfMagic: boolean;
  orderOfThePheonix: boolean;
  dumbledoresArmy: boolean;
  deathEater: boolean;
  bloodStatus: string;
  species: string;
}